driver app now checks for location service and requests it
Implemented a system to periodically check if location services are enabled.
If location services are disabled, a dialog is displayed prompting the user to enable them.
Added translations for the new dialog.
The order process for drivers now also checks for location services before starting.
Slightly modified the text in the "account not activated" dialog.