Hid FloatingActionButton in MyVehiclesView during loading and changed text in OrderView alert dialog
The FloatingActionButton in `MyVehiclesView` is now hidden when `isLoading` is true, even if `myVehicles` is empty.
The alert dialog text in `OrderView` when a driver accepts an order has been changed from "accept the order?" to "do you want to apply?".
Corresponding translations for "do you want to apply?" have been added to `locale.dart`.