Removed hours from timer, improved timer logic
The countdown timer in the order details view now only displays minutes and seconds, removing the hours component.

Additionally, the logic for checking if a timer should be displayed for driver applications has been encapsulated into a new `checkForTimer` method within the `OrderController`. This method is now called during initialization and after refreshing the order.